Factor out solarkraft home dir
